Makefile: ensure filelist-final.txt is rebuilt when needed
If one performs a build without GP tests, then extracts the GP tests
under optee_test and runs "make" again, the file initramfs.cpio.gz will
end up unchanged. This is because the file list is not updated when the
environment changes.

This patches fixes the issue by creating a file: .initramfs_exports
containing all the environment variables that influence the content of
gen_rootfs/filelist-final.txt (as well as their values). The file is
touched only when its content would change, and it is declared as a
dependancy of gen_rootfs/filelist-final.txt. This ensures that the file
list is rebuilt when its content is obsolete.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>