Makefile: fix download rules, update optee_test
- The Makefile contains obsolete rules that were intended to download
and extract the Android SDK. But, it is the NDK that is needed to build
the project. This commit fixes the rules so that the NDK is properly
retrieved and extracted. Fixes the following build error:

make[1]: *** No rule to make target `toolchains/android-ndk-r10e', needed by `build-optee-client'.  Stop.

- The image download macro (make-dl-rule) was not properly taking the
file path into account. It is fixed, too.

- The userdata image has been renamed on the Linaro snapshot server,
adjust for that too.

- This commit also fixes a wrong SHA1 reference in the optee_test
submodule (was pointing to a private commit) by updating the
repository URL in .gitmodules.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>