Add Dropbear SSH
Allows to ssh into the HiKey board. Set WITH_DROPBEAR=1 to enable.

The networking stuff (USB-Ethernet dongle support and DHCP) is
unchanged by this commit and bears the following caveats:
- Upstream kernel won't detect the dongle. Use the 96boards-hikey
kernel instead [1], and add the OP-TEE driver patch; a ready-to-use
branch is available at [2].
- Unplug any USB device before plugging the Ethernet dongle. In
particular, you can't have your PC connected to the OTG port at the
same time or the dongle won't be detected.
- DHCP needs to be started manually after boot by running:
  ./usb_net.sh

Dropbear SSH is started automatically on boot. The root password is
set to "root". Host keys are always the same.

[1] http://github.com/96boards-hikey/linux branch hikey-mainline-rebase
[2] http://github.com/jforissier/linux branch
        96boards-hikey--hikey-mainline-rebase-v4.9-rc5--optee-v14

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>